+++ src/main/java/mike/officefood/lunch/command/Command.java
@@ -1,0 +1,7 @@
+package mike.officefood.lunch.command;
+
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface Command {
+    void execute(Update update);
+}