+++ src/main/java/mike/officefood/lunch/command/CommandWithMessageService.java
@@ -1,0 +1,10 @@
+package mike.officefood.lunch.command;
+
+import mike.officefood.lunch.service.SendBotMessageService;
+import org.springframework.beans.factory.annotation.Autowired;
+
+public class CommandWithMessageService {
+    @Autowired
+    protected SendBotMessageService sendBotMessageService;
+
+}