+++ src/main/java/mike/officefood/lunch/command/UnknownCommand.java
@@ -1,0 +1,14 @@
+package mike.officefood.lunch.command;
+
+import mike.officefood.lunch.command.config.BotCommand;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+@Component
+@BotCommand(name = "/unknown")
+public class UnknownCommand extends CommandWithMessageService implements Command{
+    @Override
+    public void execute(Update update) {
+        sendBotMessageService.sendMessage(update.getMessage().getChatId().toString(), "Я не знаю этой команды:(");
+    }
+}