+++ src/main/java/mike/officefood/lunch/command/config/BotCommand.java
@@ -1,0 +1,10 @@
+package mike.officefood.lunch.command.config;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+public @interface BotCommand {
+
+    String name();
+}