+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 #Release Notes
 
+## 0.2.0-SNAPSHOT
+* added Command Pattern's logic 
+* added BotCommand Annotation and BeanPostProcessor
+
 ## 0.1.0-SNAPSHOT
 * added telegram bot
 * added SpringBoot skeleton project